refactor(backend): update QuizQuestionEntity file name typo bb-166
