+++ apps/backend/src/modules/quiz-questions/quiz-question.entity.ts
